Implement "Save as…"
Sometimes you'll want to save under another name or quickly
save a backup before editing stuff.

When opening or saving a file, we adjust the default folder and
filename of the save panel, and the window title accordingly.
This is done by adding the file's entry_ref to those messages.
update_savepanel_and_title() is called and extracts the needed
info from that entry_ref in that message.